+++ src/commands/invalid.rs
@@ -1,6 +1,6 @@
 use crate::commands::*;
 use crate::input::Input;
-use crate::output::Output;
+use crate::Output;
 
 pub fn new<TInput: Input, TOutput: Output>() -> Command<TInput, TOutput> {
     crate::commands::new::<TInput, TOutput>("invalid")
